Add handlling code for exception for findUserById, findUser, deleteUserByid in UserModel Add handling code for exception for FactoryPattern Update logic for test suit of UserModel Write some test case for Factory Pattern

+++ models/UserModel.php
@@ -30,6 +30,14 @@ class UserModel extends BaseModel
     // Get user by id:
     public function findUserById($id)
     {
+        if ($id instanceof stdClass || is_bool($id) || is_array($id) || is_null($id)) {
+            throw new InvalidArgumentException('Invalid argument');
+        }
+
+        if (!isset($id)) {
+            throw new ArgumentCountError("Too few argument");
+        }
+
         $sql = 'SELECT * FROM users WHERE id = ' . $id;
         $user = $this->select($sql);
 
@@ -39,8 +47,12 @@ class UserModel extends BaseModel
     // Get user by keyword:
     public function findUser($keyword)
     {
-        if ($keyword instanceof stdClass) {
+        if ($keyword instanceof stdClass || is_bool($keyword) || is_array($keyword) || is_null($keyword)) {
             throw new InvalidArgumentException('Invalid argument');
+        }
+
+        if (!isset($keyword)) {
+            throw new ArgumentCountError("Too few argument");
         }
 
         $sql = 'SELECT * FROM users WHERE users.name LIKE ' . '\'%' . $keyword . '%\'' . ' OR email LIKE ' . '\'%' . $keyword . '%\'';
@@ -71,6 +83,14 @@ class UserModel extends BaseModel
      */
     public function deleteUserById($id)
     {
+        if ($id instanceof stdClass || is_bool($id) || is_array($id) || is_null($id)) {
+            throw new InvalidArgumentException('Invalid argument');
+        }
+
+        if (!isset($id)) {
+            throw new ArgumentCountError("Too few argument");
+        }
+
         $sql = 'DELETE FROM users WHERE id = ' . $id;
         return $this->delete($sql);
     }
@@ -164,11 +184,13 @@ class UserModel extends BaseModel
     }
 
     // Code for testing
-    public function startTransaction(){
+    public function startTransaction()
+    {
         self::$_connection->begin_transaction();
     }
 
-    public function rollback(){
+    public function rollback()
+    {
         self::$_connection->rollback();
     }
 }
